Fix bug from previous commit

+++ FileSystem.h
@@ -2152,13 +2152,12 @@ class FileSystem {
             for (off_t i = mid; i != dataRangeCount; ++i) {
               struct DataRange* range3 = dataRanges[i];
               if (offset + length == range3->offset) {
-                off_t oldOff = offset;
                 for (off_t j = i; j != 0; --j) {
                   struct DataRange* range4 = dataRanges[j - 1];
-                  if (offset >= range4->offset &&
-                      offset < range4->offset + range4->size) {
+                  if (offset < range4->offset + range4->size &&
+                      offset + length >= range4->offset) {
+                    length += offset - range4->offset;
                     offset = range4->offset;
-                    length += oldOff - offset;
                     break;
                   }
                 }
@@ -2167,7 +2166,7 @@ class FileSystem {
                 memmove(range3->data + length, range3->data, range3->size);
                 for (off_t j = i - 1; j >= 0; --j) {
                   struct DataRange* range4 = dataRanges[j];
-                  if (offset < range4->offset + range4->size ||
+                  if (offset < range4->offset + range4->size &&
                       offset + length >= range4->offset) {
                     memmove(range3->data + (range4->offset - offset), range4->data, range4->size);
                     RemoveRange(j);
